recommender/requirements.txt
asgiref==3.7.2
black==23.11.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
Django==4.2.7
djangorestframework==3.14.0
idna==3.4
joblib==1.3.2
mypy-extensions==1.0.0
numpy==1.26.2
packaging==23.2
pandas==2.1.3
pathspec==0.11.2
platformdirs==4.0.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
